ai-content-creator/requirements.txt
python-dotenv>=1.0.1
openai>=1.60.0
pip install python-dotenv